Impl most of eval
All that's left is to impl substitute and add tests.

At first, I thought this would be a nicer way of doing things because it's sort
of more type safe than before (since things that are lists, like application,
are vecs), but now I think it's actually much more complicated than necessary.
Lifetimes are easier to deal with this way because Vec's take care of most of
the complicated stuff. But in the end, I still end up needing to clone vecs when
doing evaluation, which is far from ideal.

I think a better approach is to use a memory pool. I've thought about that
before, but I wanted to try using vec's to see if the code was nicer.

I think this code may be more "rusty" in the sense that you can more easily see
exactly what operations are being performed at each step, but it's also more
error prone (especially the parser).